+++ NEWS.md
@@ -5,7 +5,9 @@
 * Make `formatStats()` robust to the presence of character values.
 * Stop dropping zero-valued columns in `formatStats()`.
 * New `residualBootstrapTree()` function.
-* `makeMatrix()` now warns and returns NULL if a column is not found.
+* `makeMatrix()` now warns and returns `NULL` if a column is not found.
+* `makeValueTibble()` returns the original tree if `NULL` is passed instead of
+   a matrix.
 
 # ScrambledTreeBuilder 1.1.0